Refactoring. Removed some duplicate code. Refactored drawing logic into separate file. Replaced "open" statements with "with" statements, or "try/finally" in some cases.
