+++ webserver.py
@@ -4,22 +4,15 @@
 # Drawing backend web server
 #
 # Contact /u/minlite for comments/suggestions
-
-from time import gmtime, strftime, sleep
-import threading
-import json
-import hashlib
 import sys
 
 from flask import Flask, request, render_template, redirect, url_for
 from flask.ext.basicauth import BasicAuth
-import dropbox
-from websocket import create_connection
-from websocket._exceptions import WebSocketConnectionClosedException
-import requests
 import praw
 
 import configuration
+from drawing import DrawingThread
+
 
 app = Flask(__name__)
 
@@ -34,8 +27,7 @@ sys.stdout = open(configuration.LOG_FILE_NAME, "w+", 0)
 @app.route('/')
 @basic_auth.required
 def homepage():
-    auth_link = r.get_authorize_url('UniqueKey',
-                                    refreshable=True)
+    auth_link = r.get_authorize_url('UniqueKey', refreshable=True)
     return render_template("homepage.html", auth_link=auth_link)
 
 
@@ -53,7 +45,8 @@ def check_username():
     auth_link = r.get_authorize_url('UniqueKey',
                                     refreshable=True)
 
-    return render_template("checkusername.html", username=user.name, moderator=configuration.MODERATOR_USERNAME, auth_link=auth_link)
+    return render_template("checkusername.html", username=user.name, moderator=configuration.MODERATOR_USERNAME,
+                           auth_link=auth_link)
 
 
 @app.route('/select_thread')
@@ -125,179 +118,8 @@ def public_log():
 @app.route('/get_log')
 @basic_auth.required
 def get_log():
-    return open(configuration.LOG_FILE_NAME, "r").read()
-
-
-def chunksize(size, filename):
-    f = open(filename, 'rb')
-    done = 0
-    while not done:
-        chunk = f.read(size)
-        if chunk:
-            yield chunk
-        else:
-            done = 1
-    f.close()
-    return
-
-
-def sha256(filename):
-    h = hashlib.sha256()
-    for chunk in chunksize(16384, filename):
-        h.update(chunk)
-    return h.hexdigest()
-
-
-class DrawingThread(threading.Thread):
-    def __init__(self, submission_id):
-        super(DrawingThread, self).__init__()
-        self.submission_id = submission_id
-
-    def run(self):
-        print "Getting comment ids..."
-
-        print ""
-
-        comment_ids = []
-
-        while 1:
-            response = requests.get("https://www.reddit.com/comments/" + self.submission_id + ".json?sort=old")
-            thread = response.json()
-
-            if type(thread) is list:
-                break
-
-            sleep(10)
-
-        children = thread[1]['data']['children']
-
-        for child in children:
-            if child['data']['parent_id'] == "t3_" + self.submission_id:
-                if child['kind'] == 't1':
-                    comment_ids.append(child['data']['id'])
-
-                elif child['kind'] == 'more':
-                    # More comments
-                    comment_ids[len(comment_ids):] = child['data']['children']
-
-        f_comment_ids = open('comment_ids', 'w')
-
-        for comment_id in comment_ids:
-            f_comment_ids.write(comment_id + "\n")
-
-        f_comment_ids.close()
-
-        f_comment_ids = open('comment_ids', 'rb')
-
-        client = dropbox.client.DropboxClient(configuration.DROPBOX_ACCESS_TOKEN)
-        comment_ids_response = client.put_file('/comment_ids-' + strftime("%d-%b-%Y", gmtime()) + '.txt', f_comment_ids)
-
-        comment_ids_link = client.share(comment_ids_response['path'], short_url=False)
-
-        f_comment_ids.close()
-
-        print "Participants: " + str(len(comment_ids))
-        print ""
-
-        print "Comment IDs: " + comment_ids_link['url'] + " Expires at: " + comment_ids_link['expires']
-
-        print ""
-
-        print "Comment IDs SHA-256: " + sha256('comment_ids')
-
-        print ""
-
-        print "Subscribing to the bitcoin blockchain..."
-
-        print ""
-
-        block_count = 0
-
-        while block_count < configuration.WINNER_BLOCK:
-            try:
-                ws = create_connection("wss://ws.blockchain.info/inv")
-                ws.send("{\"op\":\"blocks_sub\"}")
-                print "Waiting for block #" + str(block_count + 1)
-                block = ws.recv()
-                blockj = json.loads(block)
-                print "Block Found! Hash is: " + blockj['x']['hash']
-                block_count += 1
-                ws.close()
-                print "Waiting 10 seconds for the block to settle..."
-                sleep(10)
-            except WebSocketConnectionClosedException:
-                print "Connection Closed. Trying again..."
-
-        height = blockj['x']['height']
-
-        print ""
-
-        print "Winner height is " + str(height)
-
-        print ""
-
-        if configuration.CONFIRMATION_BLOCKS != 0:
-            print "Waiting for " + str(configuration.CONFIRMATION_BLOCKS) + " confirmations..."
-            print ""
-
-        block_count = 0
-
-        while block_count < configuration.CONFIRMATION_BLOCKS:
-            try:
-                ws = create_connection("wss://ws.blockchain.info/inv")
-                ws.send("{\"op\":\"blocks_sub\"}")
-                print "Waiting for block #" + str(block_count + 1)
-                block = ws.recv()
-                blockj = json.loads(block)
-                print "Block Found! Hash is: " + blockj['x']['hash']
-                block_count += 1
-                ws.close()
-                print "Waiting 10 seconds for the block to settle..."
-                sleep(10)
-            except WebSocketConnectionClosedException:
-                print "Connection Closed. Trying again..."
-
-        if configuration.CONFIRMATION_BLOCKS != 0:
-            print ""
-            print "Confirmation completed..."
-            print ""
-
-        print "Retrieving height " + str(height)
-
-        print ""
-
-        blocks_with_height = requests.get("https://blockchain.info/block-height/" + str(height) + "?format=json").json()
-
-        for block_with_height in blocks_with_height['blocks']:
-            if block_with_height['main_chain']:
-                blockj = block_with_height
-                break
-
-        final_hash = blockj['hash']
-
-        print "Winning hash is: " + final_hash
-
-        winner_index = 1 + (int(final_hash, 16) % len(comment_ids))
-        winner_id = comment_ids[winner_index - 1]
-
-        print "Participants: " + str(len(comment_ids))
-        print "Winner index: " + str(winner_index)
-        print "Winner comment id: " + str(winner_id)
-
-        print ""
-
-        while 1:
-            comment = requests.get(
-                "https://www.reddit.com/comments/" + self.submission_id + ".json?comment=" + winner_id).json()
-
-            if type(comment) is list:
-                break
-
-            sleep(10)
-
-        winner = comment[1]['data']['children'][0]['data']['author']
-
-        print "Winner is: " + winner
+    with open(configuration.LOG_FILE_NAME, "r") as f:
+        return f.read()
 
 
 if __name__ == '__main__':
@@ -306,4 +128,5 @@ if __name__ == '__main__':
 
     r.set_oauth_app_info(configuration.REDDIT_CLIENT_ID, configuration.REDDIT_CLIENT_SECRET,
                          configuration.REDDIT_REDIRECT_URI)
+
     app.run(debug=configuration.DEBUG, port=configuration.WEB_SERVER_PORT)
